Issue 2 test inverted. Relaxed MERGEFIELD checking
Why invert issue 2 test? Because the user (I) may want to type in
mergefields manually through Ctrl+F9, which created a ton of pain for me
when I first tried to use this package, tracing the mysterious
ValueError.

Now, any truly invalid mergefields (not something that differs only by a
space or something) will cause a warning to be thrown, hopefully
informing the user.